GateController::logRequest for minimal generic logging
GateController subclasses are encouraged to implement a logRequest
method with relevant properties to track incoming notifications into
a debug or audit log.

We provide a minimal logger with the name of the service and the door
as a fallback if a controller omits one.

Currently, request handling (and so the responsibility to log) is to
the children controller, so they must call `$this->logRequest();`.