Simple feature toggle functionnality
The notifications center includes an API entry point, and will offer
a front-end notifications viewer.

We want to allow a flexible set of features. Features are stored in
the Config repository, at the app.features key.

When you develop something for the notifications center, you can
wrap it as a feature with:

if (Features::isEnabled('quux')) {
    // Your feature init code
}

Then, add a new quux key to the features array, at config/app.php.