Reporting code
We want to provide feedback about what actions have been done
to process a specific payload.

We introduce a modular evenemential approach: any job having something
to report fill an object inheriting of Action, then fire a ReportEvent
event.

ActionsReports (with the facade Report) will collect these reports.

Finally, we're able to offer something like this, for logging purpose
or to send as a response to the API client (`Report::render()`):

```
{
  "actions": [
     ...,
     ...,
     ...
  ],
  "created": 1449409799,
  "gate": "GitHub",
  "door": "Nasqueron"
}
```