Better Phabricator API requests handling
Summary:
We don't try to guess if the API reply will be an object with the
PHID as key, or an array with 0 as offset for our reply, or an object
with 0 as property.

Instead, we call `PhabricatorAPI::getFirstResult()` method.

Test Plan:
With this change, the development server fired notifications
for pending requests.

19:42:09 < Wearg> dereckson accepted D117: PHPMD: `composer phpmd`.

Reviewers: dereckson

Differential Revision: http://devcentral.nasqueron.org/D120