ActionsReport::addAction
Summary:
We want to allow to decouple the report class, a generic object, and
the events listener, specific to the report the application emits.

For this purpose, we add a new public method, ActionsReport::addAction.

Test Plan:
Send a payload and ensure the report contains actions.

Unit tests to come in D176.

Reviewers: dereckson

Differential Revision: http://devcentral.nasqueron.org/D175