Not operational broker
Summary:
If something wrong occurs at the broker level, ensure we've a nice
exception attached to our report and a 503 status code.

This ensures a full coverage of the SendMessageToBroker job, as we
test sendMessage stores the exception and report attaches it to the
event to add to the report.

Test Plan: `phpunit`

Reviewers: dereckson

Subscribers: alken-orin

Differential Revision: http://devcentral.nasqueron.org/D216