Move to config the list of event listeners
Summary:
Currently, the event listeners list were determined automatically
at runtime reading the `app/Listeners` folder.

The intent were a drop-in for plugins systems. To be coherent with
other similar lists, like the services providers or facades ones,
we move to config/app.php this list, as the listeners key.

A test will fail if a file is forgotten to be declared.

Test Plan: Run tests

Reviewers: dereckson

Subscribers: alken-orin

Differential Revision: http://devcentral.nasqueron.org/D234