Tests for GitHubGateController
Summary:
rNOTIF920690128405 added a isValidRequest method to the class,
and amended onPost and isLegitRequest methods.

We provide tests to cover the new not legit case, and the invalid
cases not provided in test shipped with the former revision.

Test Plan: Run tests and assert we have a full class coverage

Reviewers: dereckson

Subscribers: alken-orin

Differential Revision: http://devcentral.nasqueron.org/D263