GitHub push notification message for rebase
Summary:
When a branch is rebased, then pushed with -f option to GitHub,
a push payload is fired with an empty commits array.

We provide a message to handle this case.

URL comparison is used and welcome, so we keep it.

Fixes T703.

Test Plan: Run tests

Reviewers: dereckson

Maniphest Tasks: T703

Differential Revision: http://devcentral.nasqueron.org/D273