Refactor GitHubPayloadAnalyzer into payloads individual rules
Summary:
Instead of trying to parse every payload in GitHubPayloadAnalyzer,
we split the processing into several rules class, one class per type.
They are located in `app/Analyzers/GitHub/Events`.

So, if GitHub introduces a new event type, the only thing to do is
to create a new class for this event, provide sensible return values
for `getDescription` and `getLink`. The class will then be autoloaded.

Test Plan:
`phpunit`

Code coverage at 100% for the new classes and the
modified parts of GitHubPayloadAnalyzer.

Reviewers: dereckson

Subscribers: alken-orin

Differential Revision: http://devcentral.nasqueron.org/D274