Add support for Jenkins notification plug-in
Summary:
A Jenkins plugin allow to send a JSON payload to an HTTP end-point:
https://wiki.jenkins-ci.org/display/JENKINS/Notification+Plugin

Support for this payload is added on the newly created Jenkins gate.

This allows to track completed or failed Jenkins job.

In this first iteration, it will sort all the notifications to the
default "ci" group. Plan is to implement in a follow-up change the
mapping between jobs and groups.

Ref T953

Reviewers: dereckson

Subscribers: alken-orin

Maniphest Tasks: T953

Differential Revision: https://devcentral.nasqueron.org/D612