Add notifications:payload command
Summary:
A CLI command to get a notification payload from a service payload.

The goal is to ease debug of consumers applications: you can get a
sample notification without the assistance of the message broker
for a specified problematic payload.

Fixes T983.

Test Plan: Used for D347.

Reviewers: dereckson

Subscribers: alken-orin

Maniphest Tasks: T983

Differential Revision: https://devcentral.nasqueron.org/D617