Schema for Jenkins job notifications
Summary:
The schema is a JSON object containing 3 properties:

  - defaultGroup: the fallback group when a job has no mapping
  - groupsMapping: an array containing the job/group mappings
  - notifyOnlyOnFailure: an array of the jobs we ignore all events
    excepted failure.

The defaultGroup default value is "ci" per 8d700507fd7c.

The groupsMapping is similar to Phabricator ones, a collection of
objects with two properties:
  - group (string): the name of the group receiving the notification
  - jobs (array): the names of the jobs for this group (* allowed)

The notifyOnlyOnFailure is a workaround to a limitation of the Jenkins
Notifications plug-in, which doesn't have an option to notify only when
the job has failed.

Ref T953

Test Plan: Prepare a configuration file for Nasqueron jobs: P216.

Reviewers: dereckson

Maniphest Tasks: T953

Differential Revision: https://devcentral.nasqueron.org/D620