Improve documentation and code types coherence
Summary:
We use the phan static analysis tool to detect divergences
between types in the documentation and the source code.

Types given in absolute namespace notation are now prepended with \.
When the type is in the current namespace, relative notation is used.

The new PHP 7.1 pseudo-type `iterable` has been chosen to use with
PhabricatorStory, as there is no need for the payload sent by
Phabricator API to be an array.

This is no-op as only method signatures and documentation is updated.

Test Plan: `phan | grep PhanTypeMismatch`

Reviewers: dereckson

Reviewed By: dereckson

Subscribers: alken-orin

Differential Revision: https://devcentral.nasqueron.org/D769