Declare explicitely return types
Summary:
This commit allows us to consider this code base to be PHP 7.

To explicitly declare return type and non object parameters
allow us to be sure code documentation is up to date and tell
the truth, and to allow to detect regression immediately in
the future. That improves codebase quality.

Phabricator is currently left alone, as we need to refactor
this part of the code. We'll add return type as part of the
refactoring.

Fixes T1062.

Test Plan:
The code touched is the part of code base with full code coverage.
Tests was run at each iteration to detect type issues.

Reviewers: dereckson

Subscribers: alken-orin

Maniphest Tasks: T1062

Differential Revision: https://devcentral.nasqueron.org/D784