Remove GitHub to AMQP mirror feature
Summary:
Before the notifications system, we introduced a direct GitHub to AMQP broker
gateway at 7e5bd3c8b.

This system helped to implement the broker aspects, but was then superseded
by our standard notifications format.

As we strive to simplify the events propagation to ease to add more services,
and as nobody seems to use it, we remove this legacy feature.

Test Plan: Run tests. No unit test covered the feature.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D808