Prune unused User class
Summary:
We don't auth users in this application.

This is a server to receive API messages, not for users.

At first, it was imagined a client could be integrated to
this application.

A web client is probably a good idea. To host the client at
the same URL too. But none of these facts means we should
use this gateway application in this goal.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D810