Ensure ItemGroupMapping::$items is always an array
Summary:
When a JSON configuration file omits an array, we can assume
the configuration author intent is to pass an empty array.

Fixes T1179.

Test Plan:
A new unit test is provided, to ensure the payload is well
expanded with an empty array and not a null value.

The tests correctly fails before the fix and passes after:

```
-    'items' => null
+    'items' => Array ()
```

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1179

Differential Revision: https://devcentral.nasqueron.org/D947