Solve Composer conflict dependency
Summary:
Conflict:

- Phan 0.9.5 introduced a dependency to nikic/php-parser, in 3.1.1
- psysh 0.7 requires the same dependency, but in version 1.2.2-2.1.1
- We currently use Laravel 5.2, which requires psysh 0.6 or 0.7
- We need psysh 0.8+ to solve this dependency issue
- Laravel 5.3+ is happy with psysh 0.8+

Plan is so to force the last version of Phan before the dependnecy
to nikic/php-parser, ie 0.9.4.

When the code will be upgraded to use Laravel 5.3+,
to bump the Phan version will then be possible.

The composer package has been renamed to phan/phan, so this change
also switches to this new vendor name.

Fixes T1258.

Test Plan:
```
$ rm -rf vendor composer.lock
$ composer install
```

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1258

Differential Revision: https://devcentral.nasqueron.org/D1061